pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "Diff tool for comparing JSON and YAML files. Prints out differences in three formats: stylish (default; show all changed and unchanged contents and their inner structure; plain (shows only changes and hide unchanged nested contents); JSON (show all generated data in JSON format)."
authors = ["fiftinmen <fiftin.men@gmail.com>"]
classifiers = [
    "Topic :: Text Processing :: General",
    "Topic :: Text Processing :: Filters",
    "Topic :: Text Processing :: Utilities",
]
repository = "https://github.com/fiftinmen/python-project-49"
readme = "README.md"
packages = [
    {include = "gendiff"},
]

[tool.poetry.dependencies]
python = "^3.10,<4.0"
pyyaml = "^6.0.0,<7.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
pytest-cov = "^4.1.0"
flake8 = "^6.1.0"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
